+++ src/cern/colt/map/HashFunctions.java
@@ -28,7 +28,7 @@ public class HashFunctions extends Object {
      * @return a hash code value for the specified value.
      */
     public static int hash(char value) {
-        return (int) value;
+        return value;
     }
 
     /**
@@ -105,7 +105,7 @@ public class HashFunctions extends Object {
      * @return a hash code value for the specified value.
      */
     public static int hash(short value) {
-        return (int) value;
+        return value;
     }
 
     /**
