+++ src/cern/colt/matrix/tfloat/algo/solver/FloatChebyshev.java
@@ -88,7 +88,8 @@ public class FloatChebyshev extends AbstractFloatIterativeSolver {
             throw new IllegalArgumentException("eigmin > eigmax");
     }
 
-    public FloatMatrix1D solve(FloatMatrix2D A, FloatMatrix1D b, FloatMatrix1D x) throws IterativeSolverFloatNotConvergedException {
+    public FloatMatrix1D solve(FloatMatrix2D A, FloatMatrix1D b, FloatMatrix1D x)
+            throws IterativeSolverFloatNotConvergedException {
         checkSizes(A, b, x);
 
         float alpha = 0, beta = 0, c = 0, d = 0;
@@ -103,11 +104,11 @@ public class FloatChebyshev extends AbstractFloatIterativeSolver {
 
             if (iter.isFirst()) {
                 p.assign(z);
-                alpha = (float) (2.0 / d);
+                alpha = 2.0f / d;
             } else {
-                beta = (float) ((alpha * c) / 2.0);
+                beta = (alpha * c) / 2.0f;
                 beta *= beta;
-                alpha = (float) (1.0 / (d - beta));
+                alpha = 1.0f / (d - beta);
                 p.assign(z, FloatFunctions.plusMultFirst(beta));
             }
